third_party/WebKit/LayoutTests/http/tests/security/xssAuditor/anchor-url-dom-write-location-expected.txt

CONSOLE ERROR: line 6: The XSS Auditor refused to execute a script in 'http://localhost:8000/security/xssAuditor/resources/echo-dom-write-location.php?#<script>alert(String.fromCharCode(0x58,0x53,0x53))</script>' because its source code was found within the request. The server sent an 'X-XSS-Protection' header requesting this behavior.
